More RbxxString functions
Made it so that the end user can not mutate RbxxStrings and make use of certain utils like upcase and downcase.

Also will probably be making use of namespaces heavily in the future to increase readability of code.

Also thinking of making use of a ruby_t as a type alias to VALUE.

Also also also, lots of these are simplified versions of their ruby counterparts so not all variants of even these functions that were added are available such as splices and use of regular expressions and locales.

Should definitely make sure that any rbxx type creates the same exception behavior just the same as it creates the same expected return values.