Terrible commit message
Got some thoughts on what I'm gonna add, but for right now the only important new thing is the map class.
This the first templated class exposed to Ruby, and will probably cause vector to be reworked. I've also started to use some of the techniques that I picked up from making an SDL2 wrapper for Ruby in order to make an asteroids clone. That SDL2 wrapper will also more than likely end up as part of rbxx. Along with maybe a few other libraries that I use a lot in more code. Keep in mind that this is meant to become a sort of standard library for compiling Ruby into C++ and allowing Ruby and C++ to communicate more effectively.

So the current way of doing that is to create classes that act as much like their ruby equivalents as possible to take advantage of duck typing to be able to integrate the two languages more easily.
The main difference is there is static type checking built into these types so that they are not quite as malleable as their ruby equivalents. So you cannot have an rbxx::map<int, std::string> and try to set one of it's keys to be "10". Simply won't work. You have to cast/translate it first.